services/backend-fastapi/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
# Logging
loguru==0.7.2
# HTTP Client
httpx==0.25.1
# Async Support
aiofiles==23.2.1
tenacity==8.2.3
# Environment and Config
python-dotenv==1.0.0
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.31.0
# Image Processing
Pillow==10.1.0
# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3
# Utilities
python-jose[cryptography]==3.3.0  # For JWT
passlib[bcrypt]==1.7.4  # For password hashing
# Add these to your existing requirements.txt
numpy==1.26.2
scikit-learn==1.3.2  # Likely needed for quality scoring
torch==2.1.1  # If using PyTorch models
transformers==4.35.2  # For huggingface transformers

# backend-fastapi/requirements.txt additions
aiohttp>=3.8.0
pyzmq>=22.3.0

# ai-powerhouse-pipelines/requirements.txt additions
pyzmq>=22.3.0